backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
mistralai==2.3.1
asyncpg==0.30.0
pgvector==0.3.5
supabase==2.9.1
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
